Want simple performance tool
This tool allows us to pump IO full speed to supported storage URI's.
The intent of this is to be able to test performance without the need
for setting up any targets and, subsequently, benchmark tools.

Its no where near as sophisticated as fio, but thats also not the
intent. The intent is to get good numbers but also, to assist in
testing/developing the handling of IO errors and rebuild

+++ mayastor/src/bdev/nexus/nexus_bdev_children.rs
@@ -419,7 +419,7 @@ impl Nexus {
                         self.name, *s, self.bdev.alignment()
                     );
                     unsafe {
-                        (*self.bdev.as_ptr()).required_alignment = *s;
+                        (*self.bdev.as_ptr()).required_alignment = *s as u8;
                     }
                 }
             })
